+++ packages/terra-framework-docs/src/terra-dev-site/doc/collapsible-menu-view/example/CollapsibleMenuViewStartAligned.jsx
@@ -90,11 +90,12 @@ class CollapsibleMenuViewStartAligned extends React.Component {
           ]}
         />
         <CollapsibleMenuView.Divider key="Divider2" />
-        <CollapsibleMenuView.ItemGroup ariaLabel="View Type Single Selection" key="ViewTypeSelection" selectedKeys={[this.state.displayType]} onChange={this.handleDisplayTypeChange}>
+        <CollapsibleMenuView.ItemGroup ariaLabel="View Type Single Selection" groupId="view-type-single-select" key="ViewTypeSelection" selectedKeys={[this.state.displayType]} onChange={this.handleDisplayTypeChange}>
           <CollapsibleMenuView.Item
             icon={<IconTable />}
             text="Table View"
             key="tableView"
+            ariaDescribedBy="view-type-single-select"
             isIconOnly
             shouldCloseOnClick={false}
             isSelected={this.state.displayType === 'tableView'}
@@ -103,6 +104,7 @@ class CollapsibleMenuViewStartAligned extends React.Component {
             icon={<IconVisualization />}
             text="Chart View"
             key="chartView"
+            ariaDescribedBy="view-type-single-select"
             isIconOnly
             shouldCloseOnClick={false}
             isSelected={this.state.displayType === 'chartView'}
@@ -111,6 +113,7 @@ class CollapsibleMenuViewStartAligned extends React.Component {
             icon={<IconTreemap />}
             text="Treemap View"
             key="treemapView"
+            ariaDescribedBy="view-type-single-select"
             isIconOnly
             shouldCloseOnClick={false}
             isSelected={this.state.displayType === 'treemapView'}